+++ src/components/Button/Button.tsx
@@ -38,7 +38,7 @@ const buttonVariants = tv({
         // text color
         "text-gray-900 dark:text-gray-50",
         // background color
-        " bg-white dark:bg-gray-950",
+        "bg-white dark:bg-gray-950",
         //hover color
         "hover:bg-gray-50 dark:hover:bg-gray-900/60",
         // disabled