+++ src/utils/hasErrorInput.ts
@@ -1,10 +1,10 @@
 // Tremor Raw hasErrorInput [v0.0.1]
 
 export const hasErrorInput = [
-    // base
-    "ring-2",
-    // border color
-    "border-red-500 dark:border-red-700",
-    // ring color
-    "ring-red-200 dark:ring-red-700/30",
+  // base
+  "ring-2",
+  // border color
+  "border-red-500 dark:border-red-700",
+  // ring color
+  "ring-red-200 dark:ring-red-700/30",
 ]