+++ src/utils/hasOnlyOneValueForKey.ts
@@ -1,0 +1,19 @@
+// Tremor Raw hasOnlyOneValueForKey [v0.1.0]
+
+export function hasOnlyOneValueForKey(
+  array: any[],
+  keyToCheck: string,
+): boolean {
+  const val: any[] = []
+
+  for (const obj of array) {
+    if (Object.prototype.hasOwnProperty.call(obj, keyToCheck)) {
+      val.push(obj[keyToCheck])
+      if (val.length > 1) {
+        return false
+      }
+    }
+  }
+
+  return true
+}