+++ vitest.config.ts
@@ -1,0 +1,9 @@
+import tsconfigPaths from "vite-tsconfig-paths"
+import { defineConfig } from "vitest/config"
+
+export default defineConfig({
+  test: {
+    include: ["src/utils/utils.spec.ts"],
+  },
+  plugins: [tsconfigPaths()],
+})