+++ src/components/DatePicker/changelog.md
@@ -1,13 +1,19 @@
 # Tremor Raw Date Picker Changelog
 
+## 1.0.1
+
+### Changes
+
+- Fix: Add transition class to time input
+
 ## 1.0.0
 
 ### Changes
 
-* New: Add time functionality (showTimePicker)
+- New: Add time functionality (showTimePicker)
 
 ## 0.0.1
 
 ### Changes
 
-* New: Align prop to horizontally align the popover ("center" | "end" | "start")
+- New: Align prop to horizontally align the popover ("center" | "end" | "start")