+++ src/components/Input/changelog.md
@@ -1,7 +1,16 @@
 # Tremor Raw Input Changelog
 
+## 1.0.0
+
+### Changes
+
+- New: Adds an inputClassName prop to pass styles to the nested input
+- Fix: Making it full width to behave like a native input
+- Fix: Pulls className into the outer most component
+- Fix: Adds transition class
+
 ## 0.0.1
 
 ### Changes
 
-* Improved focus mode
+- Improved focus mode