+++ src/components/Label/changelog.md
@@ -1,5 +1,7 @@
 # Tremor Raw Label Changelog
 
-## 0.0.0
+## 0.0.1
 
 ### Changes
+
+- Fix: Label props name