+++ src/components/RadioGroup/changelog.md
@@ -1,5 +1,7 @@
 # Tremor Raw RadioGroup Changelog
 
-## 0.0.0
+## 0.0.1
 
 ### Changes
+
+- Fix: Remove duplicate size