+++ src/hooks/changelog.md
@@ -1,5 +1,13 @@
 # Tremor Raw hooks Changelog
 
+## 0.0.1
+
+### Changes
+
+- Fix: Maximum update depth exceeded error on useOnWindowResize
+
 ## 0.0.0
 
 ### Changes
+
+- New: useOnWindowResize hook for charts